+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvJointGridContainerSource.java
@@ -101,8 +101,7 @@ public class CsvJointGridContainerSource {
           gridName,
           rawGridElements.getOrThrow(),
           systemParticipants.getOrThrow(),
-          emUnits.getOrThrow(),
-          graphicElements.getOrThrow());
+          emUnits.getOrThrow());
     }
   }
 }